refactor: renamed maintenance components
